% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.class}
\alias{attribute.class}
\title{get the class for each column in data, as a named character vector}
\usage{
attribute.class(data, .checkAssertions = T)
}
\arguments{
\item{data}{a data frame or tibble type object}
}
\value{
a character vector of the class type of each column
}
\description{
get the class for each column in data, as a named character vector
}
\examples{
sample.data <- data.frame('chr' = sample(c('a','b','c'), 100, T),
                          'int' = sample(c(1,2,3), 100, T),
                          'dttm' = sample(as.Date.character(paste('2020-01-', 1:30, sep = '')), 100, T))

# get data classes for all columns
attribute.class(sample.data)
}
