% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.summary}
\alias{attribute.summary}
\title{generate summary for a given attribute grouped by a target attribute}
\usage{
attribute.summary(
  data,
  summary_attr = NULL,
  group_attr = NULL,
  .checkAssertions = T
)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{summary_attr}{a column label as a string to summarize}

\item{group_attr}{a column label to group by before summarizing (optional)}
}
\value{
a grouped summary table for the attribute(s)
}
\description{
generate summary for a given attribute grouped by a target attribute
}
\examples{
# summarise all columns
attribute.summary(iris)

# summarise only 'Petal.Length'
attribute.summary(iris, 'Petal.Length')

# summarise 'Petal.Length' grouped by 'Species'
attribute.summary(iris, 'Petal.Length', 'Species')

# summarise all columns grouped by 'Species'
attribute.summary(iris, group_attr = 'Species')
}
