% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.types}
\alias{attribute.types}
\title{get the type of attribute in each column of the data, as a vector}
\usage{
attribute.types(data)
}
\arguments{
\item{data}{a data frame or tibble type object}
}
\value{
a character vector of the class type of each column
}
\description{
get the type of attribute in each column of the data, as a vector
}
