% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attributes.summarise}
\alias{attributes.summarise}
\title{generate a summary of each attribute in the dataset grouped by a target attribute}
\usage{
attributes.summarise(data, group_attr = NULL, .checkAssertions = T)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{group_attr}{a column label as a string to group by before summarizing (optional)}
}
\description{
generate a summary of each attribute in the dataset grouped by a target attribute
}
\examples{
# summarise all columns
attributes.summarise(iris)

# summarise all columns grouped by 'Species'
attributes.summarise(iris, 'Species')
}
