% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{classification_metrics}
\alias{classification_metrics}
\title{extract all performance metrics for a binomial classification model}
\usage{
classification_metrics(mod, pi_0 = 0.5)
}
\arguments{
\item{mod}{a general linear model of family binomial}

\item{pi_0}{the threshold of determination used in classifaction (optional, default = 0.5)}
}
\value{
A list of classification performance metrics including the vector of predictions y.hat and the confusion matrix
}
\description{
extract all performance metrics for a binomial classification model
}
\examples{
X <- rnorm(100)
Y <- ifelse(X < 0.5, 0, 1 - sample(0:1, prob = c(0.9, .1)))
f <- glm(Y ~ X, family='binomial')
classification_metrics(f)
}
