% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{train_test_val}
\alias{train_test_val}
\title{create training, testing and validation datasets in specified proportion}
\usage{
train_test_val(data, prop_train = 0.8, prop_test = 0, prop_validation = 0)
}
\arguments{
\item{data}{a data.frame type object to split}

\item{prop_train}{proportion of data in training set}

\item{prop_test}{proportion of data in test set}

\item{prop_validation}{proportion of data in validation set (optional)}
}
\value{
a list containing the train and test data (and validation data if proportion > 0)
}
\description{
create training, testing and validation datasets in specified proportion
}
\examples{
data.ex <- data.frame('A' = sample(1:100,100), 'B' = sample(101:200,100))
train_test_val(data.ex, 0.6)
train_test_val(data.ex, 0.6, 0.3, 0.1)
}
