% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.stats}
\alias{attribute.stats}
\title{get the summary statistics for an attribute (grouped or otherwise)}
\usage{
attribute.stats(data, stats_attr)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{stats_attr}{a column label as a string to get stats for}
}
\value{
a tibble with the summary stats for the specified column/grouping
}
\description{
get the summary statistics for an attribute (grouped or otherwise)
}
