% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extradplyr.R
\name{attribute.summary}
\alias{attribute.summary}
\title{generate summary for a given attribute grouped by a target attribute}
\usage{
attribute.summary(data, summary_attr, group_attr = NULL)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{summary_attr}{a column label as a string to summarize}

\item{group_attr}{a column label to group by before summarizing (optional)}
}
\value{
a grouped summary table for the attribute
}
\description{
generate summary for a given attribute grouped by a target attribute
}
